requirements.txt
PyMySQL
discord.py